requirements.txt
apache-airflow==2.8.1
psycopg2-binary==2.9.9
pandas==2.2.2
requests==2.31.0
snowflake-connector-python[pandas]==3.15.0 
snowflake-sqlalchemy==1.5.0
pyspark==3.5.1
sqlalchemy
pyyaml
python-dotenv
pendulum
python-dotenv
phonenumbers
